+++ ripper/constants.py
@@ -16,8 +16,9 @@ LOGO_COLOR = f'''[deep_sky_blue1]
 [grey53]
 It is the end user's responsibility to obey all applicable laws.
 It is just like a server testing script and Your IP is visible.
-
 Please, make sure you are ANONYMOUS!
+
+[u blue link=https://github.com/alexmon1989/russia_ddos]https://github.com/alexmon1989/russia_ddos
 '''
 
 LOGO_NOCOLOR = f'''
@@ -31,8 +32,9 @@ LOGO_NOCOLOR = f'''
 
 It is the end user's responsibility to obey all applicable laws.
 It is just like a server testing script and Your IP is visible.
-
 Please, make sure you are ANONYMOUS!
+
+https://github.com/alexmon1989/russia_ddos
 '''
 
 BANNER = '\n\n[r][deep_sky_blue1]#StandWith[bright_yellow]Ukraine\n'
@@ -40,12 +42,7 @@ BANNER = '\n\n[r][deep_sky_blue1]#StandWith[bright_yellow]Ukraine\n'
 
 # ==== Error messages ====
 GETTING_SERVER_IP_ERROR_MSG = 'Can\'t get server IP. Packet sending failed. Check your VPN.'
-NO_SUCCESSFUL_CONNECTIONS_ERROR_VPN_MSG = 'There are no successful connections more than 2 min. ' \
-                                      'Check your VPN or change host/port.'
-NO_SUCCESSFUL_CONNECTIONS_ERROR_PROXY_MSG = 'There are no successful connections more than 2 min. ' \
-                                      'Looking for successful proxies.'
 YOUR_IP_WAS_CHANGED = 'Your IP was changed!!! Check VPN connection.'
-CANNOT_SEND_REQUEST_ERR_MSG = 'Cannot send Request or Packet. Host does not respond.'
 NO_MORE_PROXIES_MSG = 'There are no more operational proxies to work with host.'
 DEFAULT_CURRENT_IP_VALUE = '...detecting'
 HOST_IN_PROGRESS_STATUS = 'HOST_IN_PROGRESS'
@@ -71,7 +68,7 @@ ARGS_DEFAULT_PROXY_TYPE = 'socks5'
 
 
 # ==== Defaults ====
-GEOIP_NOT_DEFINED = 'NOT DEFINED'
+GEOIP_NOT_DEFINED = '--'
 CONNECT_TO_HOST_MAX_RETRY = 5
 MIN_SCREEN_WIDTH = 90
 MIN_UPDATE_HOST_STATUSES_TIMEOUT = 120